Add a contactTraces node to PatientNode
