+++ server/graphql/programs/src/lib.rs
@@ -5,6 +5,9 @@ use graphql_core::pagination::PaginationInput;
 use graphql_core::standard_graphql_error::validate_auth;
 use graphql_core::standard_graphql_error::StandardGraphqlError;
 use graphql_core::ContextExt;
+use graphql_types::types::contact_trace::ContactTraceFilterInput;
+use graphql_types::types::contact_trace::ContactTraceResponse;
+use graphql_types::types::contact_trace::ContactTraceSortInput;
 use graphql_types::types::document::DocumentNode;
 use graphql_types::types::encounter::EncounterFilterInput;
 use graphql_types::types::encounter::EncounterSortInput;
@@ -54,9 +57,6 @@ use mutations::undelete_document::UndeleteDocumentInput;
 use mutations::undelete_document::UndeleteDocumentResponse;
 use mutations::update_document::*;
 use queries::contact_trace::contact_traces;
-use queries::contact_trace::ContactTraceFilterInput;
-use queries::contact_trace::ContactTraceResponse;
-use queries::contact_trace::ContactTraceSortInput;
 use service::auth::Resource;
 use service::auth::ResourceAccessRequest;
 use service::programs::patient::patient_search_central;
