+++ server/graphql/programs/src/mutations/contact_trace/update.rs
@@ -3,13 +3,12 @@ use graphql_core::{
     standard_graphql_error::{validate_auth, StandardGraphqlError},
     ContextExt,
 };
+use graphql_types::types::contact_trace::ContactTraceNode;
 use repository::{contact_trace::ContactTraceFilter, StringFilter};
 use service::{
     auth::{Resource, ResourceAccessRequest},
     programs::contact_trace::upsert::{UpsertContactTrace, UpsertContactTraceError},
 };
-
-use crate::queries::contact_trace::ContactTraceNode;
 
 #[derive(InputObject)]
 pub struct UpdateContactTraceInput {
