+++ server/graphql/programs/src/mutations/patient/insert.rs
@@ -4,8 +4,7 @@ use graphql_core::{
     standard_graphql_error::{validate_auth, StandardGraphqlError},
     ContextExt,
 };
-use graphql_general::GenderInput;
-use graphql_types::types::patient::PatientNode;
+use graphql_types::types::{patient::PatientNode, GenderInput};
 use repository::NameType;
 use service::{
     auth::{Resource, ResourceAccessRequest},
