+++ server/graphql/programs/src/mutations/patient/update.rs
@@ -4,8 +4,7 @@ use graphql_core::{
     standard_graphql_error::{validate_auth, StandardGraphqlError},
     ContextExt,
 };
-use graphql_general::GenderInput;
-use graphql_types::types::patient::PatientNode;
+use graphql_types::types::{patient::PatientNode, GenderInput};
 use service::{
     auth::{Resource, ResourceAccessRequest},
     programs::patient::{UpdatePatient, UpdatePatientError},
