+++ server/graphql/programs/src/queries/mod.rs
@@ -21,4 +21,3 @@ pub use self::encounter_fields::*;
 pub mod document_registry;
 pub use self::document_registry::*;
 pub mod contact_trace;
-pub mod program_node;
