+++ server/graphql/types/src/types/name.rs
@@ -10,6 +10,45 @@ use super::StoreNode;
 
 #[derive(Enum, Copy, Clone, PartialEq, Eq, Debug, Serialize)]
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")] // only needed to be comparable in tests
+pub enum GenderInput {
+    Female,
+    Male,
+    TransgenderMale,
+    TransgenderMaleHormone,
+    TransgenderMaleSurgical,
+    TransgenderFemale,
+    TransgenderFemaleHormone,
+    TransgenderFemaleSurgical,
+    Unknown,
+    NonBinary,
+}
+
+impl GenderInput {
+    pub fn to_domain(self) -> Gender {
+        match self {
+            GenderInput::Female => Gender::Female,
+            GenderInput::Male => Gender::Male,
+            GenderInput::TransgenderMale => Gender::TransgenderMale,
+            GenderInput::TransgenderMaleHormone => Gender::TransgenderMaleHormone,
+            GenderInput::TransgenderMaleSurgical => Gender::TransgenderMaleSurgical,
+            GenderInput::TransgenderFemale => Gender::TransgenderFemale,
+            GenderInput::TransgenderFemaleHormone => Gender::TransgenderFemaleHormone,
+            GenderInput::TransgenderFemaleSurgical => Gender::TransgenderFemaleSurgical,
+            GenderInput::Unknown => Gender::Unknown,
+            GenderInput::NonBinary => Gender::NonBinary,
+        }
+    }
+}
+
+#[derive(InputObject, Clone)]
+pub struct EqualFilterGenderInput {
+    pub equal_to: Option<GenderInput>,
+    pub equal_any: Option<Vec<GenderInput>>,
+    pub not_equal_to: Option<GenderInput>,
+}
+
+#[derive(Enum, Copy, Clone, PartialEq, Eq, Debug, Serialize)]
+#[serde(rename_all = "SCREAMING_SNAKE_CASE")] // only needed to be comparable in tests
 pub enum NameNodeType {
     Facility,
     Patient,
